add a if to display chart only on dashboard page
