+++ app/javascript/packs/application.js
@@ -25,7 +25,7 @@ import {initSlick} from '../plugins/init_slick';
 import { loadDynamicBannerText } from '../components/banner';
 import { showLesson } from '../components/show_lesson'
 import { bookMouseover } from '../components/book_mouseover';
-import { testChart } from '../components/chart';
+import { bookingChart } from '../components/booking_chart';
 
 document.addEventListener('turbolinks:load', () => {
   loadDynamicBannerText();
@@ -33,5 +33,5 @@ document.addEventListener('turbolinks:load', () => {
   initSlick();
   bookMouseover();
   showLesson();
-  testChart();
+  bookingChart();
 });
